Suppress onIsMenuChanged runtime error messages
PageRowGlobalToolBarUI makes a connection to onIsMenuChanged in
applicationWindow().globalDrawer, but somteimes it's not actually a
GLobalDrawer and doesn't have the right signal. So we want to make sure
we're only targeting it if it is in fact a GlobalDrawer. Otherwise it
won't have the needed signal and we'll get a warning.

BUG: 471561
FIXED-IN: 6.0