Proper icon tinting
kiconloader supports internal icon tinting, but we used to do that
only if and hardcoded color property was set on it.
internally it tints with system colors, which usually happen to
correspond with Kirigami.Theme colors, but if the theme chanes
(ie in plasma is not the system color/palette, or a diffenent colorSet might
have been set) then the icon color didn't follow.
We already have all the needed pieces, always recolor with a color from
the kirigami theme, unless an hardcoded color property has been set