Introduce the behavior of Plasma::IconItem roundToIconSize
Usually icon items are set by the developer to be a standard icon size,
but it not always necessary or possible, as many times the simplest
layouting is the icon item going bigger with the icon centered. This is
the case for icons used as Widget CompactRepresentation items which have
to support the panel they live on being an arbitrary size.

This commit adds a roundToIconSize property (defaulting to true) like
PlasmaCore.IconItem had. When set, it rounds icons to the standard icon
sizes, diminishing the chance of a blurry icon by accident.