Drop AbstractChip
We already have a template to describe the behaviour for chips, so
having the intermediate AbstractChip does not add a whole lot, it only
sets the default background. We can easily move that to Chip itself,
which still allows customisation to the same level as AbstractChip since
its contentItem can be replaced.