Dialog: Add API to fetch custom footer buttons created from actions
This is complementary to standardButton() API. Dialog is expected to
create only one delegate instance per action, so it is safe assume
exactly one button will be created and available to fetch for further
customizations by consumers.

In addition to the new API, this patch also fixes an issue for
DialogButtonBox whose ListView would've otherwise allocated empty space
for invisible buttons (caused by a null in the list of actions or by
an action's visibility property set to false).