MnemonicAttached: Rework event filter
Currently, every single attached property instance would install an event
filter on the window. Instead, use a singleton and install it on qGuiApp.

This is also in line with what Breeze style does. Installing it on qGuiApp
also saves us the trouble with render windows. It seems the event handling
is pretty wonky in complicated nested window setups like System Settings,
where the render window isn't determined in time.