Units: Round up font height for gridUnit instead of silently truncating it
Implicit coercion rules of C++ allowed this truncation to happen
silently without warnings. We should be explicit about it, and also
make sure that gridUnit is large enough to accommodate a text label
without causing vertical clipping.

Note that with default Noto Sans font metrics height is 17.703125, while
Label with that font has an implicit height of 18px.