InlineViewHeader: Emulate OverlayHeader using Transform
There is no GridView.headerPositioning, so a GridView header always
scrolls away, cf. QTBUG-117035

Use a Transform and scroll the headerItem together with the view,
to make sure the header sticks.

Conveniently, through GridView attached property this will only
be done when in a GridView (obviously) and when it is the headerItem's root,
not if you put it in a wrapper (then you're on your own).