ToolBar: Attempts at fixing heading size for specific fonts
Qt 6.3 introduced new property to fill the gap between Text/Label
implicitWidth and TextMetrics width/advanceWidth: renderType. It was
supposed to fix all the inconsistency issues with fonts like Noto Sans
8pt that causes our Kirigami headers to elide into oblivion.

Unfortunately though, it does not seem to be enough, as TextMetrics
still reports e.g. 36.40625 px for a "Chips" string for which a Heading
would require 38.984375 px of space. But at least now we can say we've
done everything we could on our side.

See also:
- https://bugreports.qt.io/browse/QTBUG-85936
- https://bugreports.qt.io/browse/QTBUG-83936
- https://bugreports.qt.io/browse/QTBUG-94023

CCBUG: 474881