Restore push behavior, unbreak existing apps
The behavior of PageRow is alike a browser history of undo stack: the
top of the stack is considered the page the app is currently in, rather
than the total depth.
This logic broke in the refactor, causing many apps like multi level
KCMS or discover to do double pushes keeping creating new pages in the
row.
this does the minimal indexes fix to make the existing users behave
correctly.

BUG: 474542