AboutPage,AboutItem: Fix and rework avatar loading controller
It was completely broken because of a race between Repeater::count and
actual items being available as children to iterate over. Regardless
that was an inside-out twisted data flow, and should be replaced by a
simple top-level property with a straightforward JavaScript array
aggregator and a little helper function which is reused in person
delegate.

A new exposed property provides developers with means to control the
defaults.

Potential area for improvement: both old and new code only detects
avatar URLs of authors, but it should probably also account for credits
and translators?