NavigationTab{Bar,Button}: Don't mask icons by default and remove API
Doing this is generally harmful, as it prevents the app developer from
getting the icon they asked for if it happens to have any colors other
than the base color. It's weird and confusing and works around icon
theme issues in a way that shouldn't be encouraged. We already removed
the automatic behavior in bdb322527fbc910ce7124a45dc43dfaa0fad1522 from
Kirigami.Icon and fixed many bug reports in the process.

This commit disables the behavior and removes the API to control it.

Fixes #69