GlobalToolBar: Remove superfluous piece of background
It used to be stretched vertically, but that started conflicting with
vertical centering anchors, so in attempts to fix that it got broken
and nothing is there to give it any height anymore.

Other than that, Kirigami.Theme.backgroundColor should be a default
implementation of AbstractApplicationHeader::background on any theme,
so there is no need to repeat it here, especially with such custom size.

Besides, this binding on width is 100% broken on RTL anyway.