Declare the Kirigami QML plugin as non-optional
It most certainly isn't, it contains non-trivial code that we definitely
need loaded. And on Android we already link against the Kirigami library,
so this actually makes a difference there.

This fixes various runtime issues on Android due to the plugin not being
loaded and only the Kirigami types defined in QML being available.