Page: Fix page header disappearing under the global header
With Qt 6.7 Controls' page header position is enforced and we can no
longer move it below the global header. So instead, we need to trick the
header control into placing its content below the global header by
overriding the top padding. This generally works when using ToolBar as
header type, but isn't as reliable as the original approach. Longer term
we should either fix QTBUG-120197 upstream or not make the global
toolbar part of the page.