Move Global toolbar from Page to ColumnView
The global toolbar used to be contained in Kirigami.Page itself, and
with some hacks we changed anchors of the header and contentitem in
order to move the page header in the bottom of the global toolbar.

This was a very hacky implementation and got broken in Qt 6.7 where
the page header is now always enforced to be on top of the page.

since this is global outside the page, move the item outside the page
but as an header managed by ColumnView itself

The header won't be present anymore when the page is used outside a ColumnView or PageRow (for instance in a StackView) so existing apps should be tested against this