Show a checkbox in the corner for checkable cards
Cards can be checkable. However, when they're checkable, nothing
visually suggests this, so it's not clear to the user.

To fix this, we can add a little checkbox to the corner of checkable
cards. To make that happen, we have to add the functionality to
BannerImage, since that's what draws the header. In BannerImage, it's
made an optional and off-by-default thing, and Card opts into it.