Actually close drawers on Escape
Right now clicking on the handle will give it focus, and it will eat
keyboard events, breaking the CloseOnEscape policy, until focus is
manually transferred to something inside the drawer.
If thedrawer close policy has CloseOnEscape, manage the esc event in the
handle as well to close the drawer.

An alternative approach might be actually transferring focus after
drawer open, which it might be more natural, depending what the
preferred ux is: right now, after clickign the handle, repeatedly
hitting spacebar will open and close the drawer which can be seen as a
good thing, tough if the drawer has item that accept keyboard input
would also be good transferring the focus to for instance a search field
it might contain.

https://invent.kde.org/frameworks/kirigami/-/issues/72