Adopt ContextualHelpButton
ContextualHelpButton is great! We use it all over the place in System
Settings KCMs. It's proven effective and popular.

However this isn't the only place it makes sense to use it; apps could
benefit too. But right now it lives in the kcmutils framework under the
KCM namespace, which doesn't make sense to use in apps.

Accordingly, let's lift it out of there and stick it in Kirigami!

Code is copied from the kcmutils version with a few fixups and added
documentation.