padding: make sure to polish on completion
this is incredibly important as without polish run we'll not have an
implicit size and without that our parents will not know what to do with
us. this is doubly problematic when used as a listview delegate (such as
via AbstractCard) because the cacheBuffer determines how many delegates
to create based on the size of the first delegate right after creation.
if we have no size, the listview will create hundreds of delegates,
potentially needlessly.

BUG: 477978