icon: Remove the node with lowest opacity when animation end
When using the software rendering backend for the scene graph, we may
end up in a state where we have added the nodes for animation but the
bottom-most node isn't the most transparent one. Removing the
bottom-most node then ends up confusing the state handling leading to
misplaced icons.

This is a workaround for QTBUG-123799