Padding: Remove old overridden contentItem from the visual hierarchy
This is similar to what QQC2.Control does, except we can't reliably
reset visibility (which is also transient). So let's only unset the
parent, since we set it for a new item already.

Fixes old items floating around parented to Padding item overlapping
with a new content item and not being properly sized anymore.