Dialog: Expand past preferred size to implicit size, as documented
The documentation of preferred{Height,Width} says:

> If the content, header or footer require more space, then the
> [size] of the dialog will expand to the necessary amount of space.

However, in practice, a Dialog with preferredWidth set to a given
value would cut off a content item with a wider implicitWidth,
despite it not exceeding maximum width.

Instead of choosing preferred size or implicit size based on
whether preferred size is set, pick the larger one instead.
This will work fine with the default of preferred{Height,Width} == -1
and auto-expand as advertised.