PageRow: Fix parent of Component-based pages
By using a non-graphical item we were leaving the page outside of the scene
and were getting a warning such as:
QML FindDevicesPage: Created graphical object was not placed in the graphics
scene.

This made the pages misregister in the accessibility infrastructure too,
making certain elements not be adequately registered in there.

BUG: 482753