tst_pagepool: Fix test properly
The previous attempt was "fixing" the test without understanding the
underlying issue. A change to QList value type properties handling in
QML caused this test to iterate over a mutable array-like structure
which holds a reference to its originating object and it changes every
time the underlying property changes.

We still need to iterate over all the items here, so we make a shallow
immutable copy with the usual Array.prototype.slice call beforehand.