Remove linkActivated/linkHovered from Delegate types
We want to keep the API of these as minimally divergent from the
upstream types as possible, only adding API that really hard to
replicate and is used by pretty much all users. Link activation signals
are neither, as they can be easily accessed on the content item with a
Connections object and most of our current users are not using it.

This is a partial revert of 4fb8b8490e46d223d51389746f335a6f77ae5c67