TitleSubtitle Title Accessiblity
Ignore the label item for the purposes of accessibility.

This is another one coming up for the desire to make this work with `FormCard` (see https://invent.kde.org/libraries/kirigami-addons/-/merge_requests/229#note_912858). This obviously means that this would only work with components that have a text property which is the primary use case for this in `ItemDelegate` (it'll also work in anything `AbstractButton` based). This would mean that it would only work in a `Control` based component if text was re implemented.

We need to do something here anyway as we don't want the title appearing in accessibility interfaces twice, other options would be to expose as a property or give the root component to this similar to `SubtitleContentItem` in Kirigami-addons Delegates.