Card: emit toggled() when the user clicks on the header checkbox
Checkable cards display a checkbox in the header. When clicked, they
emit toggled() which gets bubbled up to the card. However the card
doesn't emit the toggled() signal itself, instead only setting the
value of its "checked" property.

As a result, anything that's listening to the card for a toggled()
signal will not get one. Such as, for example, the screen chooser
portal code.

Fix this by emitting the signal.

BUG: 481461
FIXED-IN: 6.3