Dialog: Add some null safety
The top-level component must not assume that its parent is always
present and valid.

The previous attempt in 9332c1183e209eaf305d2f33d7fc0cd953364eab was
almost perfect, but it messed up with the top level component's parent
binding, which was never not correct to bind to a Popup's own
Overlay.overlay attached property so it caused regressions and the
patch got reverted as a whole.