Dialog: Fix ScrollView::contentItem management
ScrollView is expected to always report its contentItem as some kind of
Flickable. But an extra check wouldn't hurt for QML compiler and type
hints. Since the check was already there, let's lift it up a bit, so
that it covers both assignments. Also, that qualified property access
isn't needed inside the Qt.binding() statement, and the statement
itself can be simplified to an expression.