ContextualHelpButton: make text always accessible
ContextualHelpButton pops up a tooltip when activated. This is great for
sighted users and screen reader users that have "Present Tooltips"
enabled, but leaves the information completely inaccessible to screen
reader users with this setting turned off.

This change puts the text that would be shown as the tooltip in the
Accessible.description property of the button; this way it will always
be read out.