Dialog: Change default horizontal padding to largeSpacing
We want content to look correctly padded by default, as the zero padded
look only works for content that is expected to be scrollable or
full-width. If you have dialog content that's normal controls,
everything sticks to the edge which looks extremely ugly.

This changes the default horizontal padding to horizontalSpacing, but
does not touch the verticalPadding. It's still possible to override the
padding to zero like before, and the only difference is if the user of
Dialog does not explicitly set padding.