[icon] Fix icon colors when using Plasma platformtheme and QIcon source
When using a name as source we load the icon from the platform theme and
therefore apply the right theme colors.

However when we receive a QIcon as source it might not have been loaded
with the right theme colors.

To fix this check if the icon was obtained by name and reload it from the
platformtheme then. The late PlasmaCore.IconItem did something similar.

BUG: 491274