PrivateActionToolButton: Replace onVisibleChanged with Connections
Apparently adding any `onVisibleChanged` handler to
PrivateActionToolButton breaks the assignment of an action to the
button. Replacing the visible binding with an explicit Binding object
or using a Connections object fixes it. Since we want to be able to
override the visible binding, use a Connections object.

I have *no* idea why this happens, or why the fix works. There seems to
be something incredibly screwy going on in the QML engine here.